walo

+++ libft.h
@@ -16,11 +16,12 @@
 # include <string.h>
 # include <stdio.h>
 # include <unistd.h>
+# include <stdint.h>
 
 int				ft_atoi(const char *str);
 void			ft_bzero(void *s, size_t n);
 void			ft_bzero(void *s, size_t n);
-void			*ft_calloc(unsigned int nmemb, unsigned int size);
+void			*ft_calloc(size_t nmemb, size_t size);
 int				ft_isalnum(int c);
 int				ft_isalpha(int c);
 int				ft_isascii(int c);
